requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
gunicorn==21.2.0

# RAG System Dependencies
anthropic>=0.40.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
